// shared/types.d.ts
export interface Category {
    id: string;
    name: string;
    display_name_tr: string;
    type: string;
    description: string;
    image_url: string;
    styles: string[];
    is_active: boolean;
    created_at: string;
    updated_at: string;
}